requirements.txt
alpaca-py
configparser
pandas